requirements.txt
# Core Python dependencies
# Python 3.12+ required (specified in workflow, not as a package)

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Scientific Computing
numpy==1.26.2
scipy==1.11.4
sympy==1.12
pandas==2.1.3

# Machine Learning
torch==2.2.0  # Updated for Python 3.12 compatibility
scikit-learn==1.3.2
gpytorch==1.11  # Gaussian Processes

# Chemistry & Materials (moved to pyproject.toml[project.optional-dependencies.chem])
# Install with: pip install ".[chem]"
# pyscf==2.3.0
# rdkit==2023.9.1
# ase==3.22.1  # Atomistic Simulation Environment

# Data & Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.43
alembic==1.12.1

# Graph & Planning
networkx==3.2.1

# Units & Physics
pint==0.22

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.92.1  # Property-based testing

# Code Quality
mypy==1.7.1
black==25.9.0
ruff==0.1.6

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0

# NLP (for explanations)
sentence-transformers==2.2.2

# Cryptography (for audit integrity)
cryptography==41.0.7

# WebSockets (for real-time updates)
websockets==12.0

# SHAP (explainability)
shap==0.43.0

# BoTorch (Bayesian optimization)
botorch==0.9.5

hydra-core==1.3.2 # Configuration management
PyYAML==6.0.3 # YAML parsing
pyarrow==14.0.1 # Parquet support (12.x requires building from source on Py3.12)
omegaconf==2.3.0 # Hydra config objects
